Second Commit - KMP implementation with iOS native UI and keeping business logic complete kotlin and shared.

+++ shared/src/commonMain/kotlin/com/app/koshpal/core/data/local/Converters.kt
@@ -29,7 +29,7 @@ object Converters {
 
     @TypeConverter
     fun fromLocalDate(date: LocalDate): Long =
-        date.toEpochDays().toLong()
+        date.toEpochDays()
 
     @TypeConverter
     fun toLocalDate(epochDay: Long): LocalDate =
